+++ Objects/Objects/Structural/CSI/Analysis/CSIStories.cs
