+++ Objects/Objects/Structural/GSA/Loading/GSAPolyline.cs
